Refactor: python, cts: Move RemoteMigrate into the _cts module.
No code changes.
